Fix pre-tag review findings: long flush counts, tenantless gating, identity validation
- RowDispatcherFlushResult counts are long, matching the long-counter
  policy everywhere else in 0.4.0 and avoiding a breaking change later.
- The worker's tenantless pass runs even when the tenant source returns
  no tenants: a tenantless-only deployment legitimately has none, and
  gating retention on tenant presence retained shared data forever. The
  empty-source warning now fires only when tenanted entities exist.
- AddRowHandler rejects a repeat registration of the same pair with a
  different phase or identity (TryAddEnumerable silently kept the first)
  and rejects two handlers sharing an identity for one entity (their
  queued rows would dispatch interchangeably).

+++ Cohort/Application/IRetentionRowDispatcher.cs
@@ -20,7 +20,7 @@ public interface IRetentionRowDispatcher
 /// counts rows the flush could not dispatch (deferred-phase rows whose sweep has not
 /// settled, or rows queued behind an in-flight sibling).
 /// </summary>
-public sealed record RowDispatcherFlushResult(int InFlightRemaining, int PendingRemaining)
+public sealed record RowDispatcherFlushResult(long InFlightRemaining, long PendingRemaining)
 {
     public bool Settled => InFlightRemaining == 0 && PendingRemaining == 0;
 }
